requirements.txt
# debugpy
python-dotenv

redis

telethon
fastapi
uvicorn[standard]